Add fallback bundle method
