+++ components/nimbus/ios/Nimbus/Collections+.swift
@@ -4,10 +4,6 @@
 
 import Foundation
 
-#if canImport(UIKit)
-    import UIKit
-#endif
-
 public extension Dictionary {
     func mapKeysNotNull<K1>(_ transform: (Key) -> K1?) -> [K1: Value] {
         let transformed: [(K1, Value)] = compactMap { k, v in
@@ -38,62 +34,6 @@ public extension Dictionary {
         }
 
         return merging(defaults, uniquingKeysWith: valueMerger)
-    }
-}
-
-public extension Array where Element == Bundle {
-    /// Search through the resource bundles looking for an image of the given name.
-    ///
-    /// If no image is found in any of the `resourceBundles`, then the `nil` is returned.
-    func getImage(named name: String) -> UIImage? {
-        for bundle in self {
-            if let image = UIImage(named: name, in: bundle, compatibleWith: nil) {
-                return image
-            }
-        }
-        return nil
-    }
-
-    /// Search through the resource bundles looking for an image of the given name.
-    ///
-    /// If no image is found in any of the `resourceBundles`, then a fatal error is
-    /// thrown. This method is only intended for use with hard coded default images
-    /// when other images have been omitted or are missing.
-    ///
-    /// The two ways of fixing this would be to provide the image as its named in the `.fml.yaml`
-    /// file or to change the name of the image in the FML file.
-    func getImageNotNull(named name: String) -> UIImage {
-        guard let image = getImage(named: name) else {
-            fatalError(
-                "An image named \"\(name)\" has been named in a `.fml.yaml` file, but is missing from the asset bundle")
-        }
-        return image
-    }
-
-    /// Search through the resource bundles looking for localized strings with the given name.
-    /// If the `name` contains exactly one slash, it is split up and the first part of the string is used
-    /// as the `tableName` and the second the `key` in localized string lookup.
-    /// If no string is found in any of the `resourceBundles`, then the `name` is passed back unmodified.
-    func getString(named name: String) -> String? {
-        let parts = name.split(separator: "/", maxSplits: 1, omittingEmptySubsequences: true).map { String($0) }
-        let key: String
-        let tableName: String?
-        switch parts.count {
-        case 2:
-            tableName = parts[0]
-            key = parts[1]
-        default:
-            tableName = nil
-            key = name
-        }
-
-        for bundle in self {
-            let value = bundle.localizedString(forKey: key, value: nil, table: tableName)
-            if value != key {
-                return value
-            }
-        }
-        return nil
     }
 }
